+++ internal/lineItem/lineItem.go
@@ -2,18 +2,18 @@ package lineItem
 
 type LineItem struct {
 	item  string
-	total float64
+	total int
 }
 
-func NewLineItem(anItem string, aTotal float64) LineItem {
+func NewLineItem(anItem string, aTotal int) LineItem {
 	return LineItem{item: anItem, total: aTotal}
 }
 
-func (li *LineItem) Total() float64 {
+func (li *LineItem) Total() int {
 	return li.total
 }
 
-func (li *LineItem) AddToPurchaseMap(aListOfPurchases *map[string]float64) {
+func (li *LineItem) AddToPurchaseMap(aListOfPurchases *map[string]int) {
 	if _, ok := (*aListOfPurchases)[li.item]; !ok {
 		(*aListOfPurchases)[li.item] = 0
 	}
