+++ internal/sale/sale.go
@@ -14,7 +14,7 @@ func NewSale(aTicket ticket.Ticket, aUser *userCredentials.UserCredentials) Sale
 	return Sale{ticket: aTicket, owner: aUser}
 }
 
-func (s *Sale) AddToPurchasesIfOwnerIs(aUser *userCredentials.UserCredentials, aListOfPurchases *map[string]float64) {
+func (s *Sale) AddToPurchasesIfOwnerIs(aUser *userCredentials.UserCredentials, aListOfPurchases *map[string]int) {
 	if aUser.SameCredentialsAs(s.owner) {
 		s.ticket.AddLineItemsAndItsCostToMapCollector(aListOfPurchases)
 	}
